# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FlowDesk API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardStats {
  totalCommands: number;
  executionsToday: number;
  executionsThisWeek: number;
  successRate: number;
  avgExecutionTime: number;
  activeWorkflows: number;
  clipboardItems: number;
  uptime: string;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  command_executed: "command_executed",
  command_created: "command_created",
  command_updated: "command_updated",
  shortcut_triggered: "shortcut_triggered",
  clipboard_copy: "clipboard_copy",
  settings_changed: "settings_changed",
} as const;

export type ActivityItemStatus =
  (typeof ActivityItemStatus)[keyof typeof ActivityItemStatus];

export const ActivityItemStatus = {
  success: "success",
  error: "error",
  warning: "warning",
  info: "info",
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  title: string;
  description?: string;
  status: ActivityItemStatus;
  timestamp: string;
  commandId?: string;
}

export interface ChartDataPoint {
  date: string;
  executions: number;
  errors: number;
  avgTime: number;
}

export type CommandStatus = (typeof CommandStatus)[keyof typeof CommandStatus];

export const CommandStatus = {
  active: "active",
  inactive: "inactive",
  draft: "draft",
} as const;

export interface Command {
  id: string;
  name: string;
  description?: string;
  category: string;
  trigger?: string;
  script?: string;
  shortcut?: string;
  status: CommandStatus;
  executionCount: number;
  lastExecuted?: string;
  avgExecutionTime?: number;
  tags?: string[];
  createdAt: string;
  updatedAt?: string;
}

export interface CreateCommandBody {
  name: string;
  description?: string;
  category: string;
  trigger?: string;
  script?: string;
  shortcut?: string;
  tags?: string[];
}

export type UpdateCommandBodyStatus =
  (typeof UpdateCommandBodyStatus)[keyof typeof UpdateCommandBodyStatus];

export const UpdateCommandBodyStatus = {
  active: "active",
  inactive: "inactive",
  draft: "draft",
} as const;

export interface UpdateCommandBody {
  name?: string;
  description?: string;
  category?: string;
  trigger?: string;
  script?: string;
  shortcut?: string;
  status?: UpdateCommandBodyStatus;
  tags?: string[];
}

export type ExecutionResultStatus =
  (typeof ExecutionResultStatus)[keyof typeof ExecutionResultStatus];

export const ExecutionResultStatus = {
  success: "success",
  error: "error",
} as const;

export interface ExecutionResult {
  id: string;
  commandId: string;
  status: ExecutionResultStatus;
  output?: string;
  error?: string;
  executionTime: number;
  executedAt: string;
}

export type HistoryEntryStatus =
  (typeof HistoryEntryStatus)[keyof typeof HistoryEntryStatus];

export const HistoryEntryStatus = {
  success: "success",
  error: "error",
  warning: "warning",
} as const;

export type HistoryEntryTriggeredBy =
  (typeof HistoryEntryTriggeredBy)[keyof typeof HistoryEntryTriggeredBy];

export const HistoryEntryTriggeredBy = {
  manual: "manual",
  shortcut: "shortcut",
  schedule: "schedule",
  api: "api",
} as const;

export interface HistoryEntry {
  id: string;
  commandId?: string;
  commandName?: string;
  type: string;
  status: HistoryEntryStatus;
  output?: string;
  error?: string;
  executionTime: number;
  triggeredBy: HistoryEntryTriggeredBy;
  timestamp: string;
}

export interface HistoryResponse {
  entries: HistoryEntry[];
  total: number;
  page: number;
  totalPages: number;
}

export interface HistorySummary {
  totalExecutions: number;
  successCount: number;
  errorCount: number;
  warningCount: number;
  mostUsedCommand?: string;
  peakHour?: string;
}

export type SettingsTheme = (typeof SettingsTheme)[keyof typeof SettingsTheme];

export const SettingsTheme = {
  dark: "dark",
  light: "light",
  system: "system",
} as const;

export interface Settings {
  theme: SettingsTheme;
  language?: string;
  notifications: boolean;
  autoStart: boolean;
  minimizeToTray?: boolean;
  showInDock?: boolean;
  globalShortcutEnabled: boolean;
  clipboardHistoryLimit?: number;
  defaultCategory?: string;
  analyticsEnabled?: boolean;
}

export type UpdateSettingsBodyTheme =
  (typeof UpdateSettingsBodyTheme)[keyof typeof UpdateSettingsBodyTheme];

export const UpdateSettingsBodyTheme = {
  dark: "dark",
  light: "light",
  system: "system",
} as const;

export interface UpdateSettingsBody {
  theme?: UpdateSettingsBodyTheme;
  language?: string;
  notifications?: boolean;
  autoStart?: boolean;
  minimizeToTray?: boolean;
  showInDock?: boolean;
  globalShortcutEnabled?: boolean;
  clipboardHistoryLimit?: number;
  defaultCategory?: string;
  analyticsEnabled?: boolean;
}

export interface Shortcut {
  id: string;
  name: string;
  description?: string;
  keys: string;
  action: string;
  enabled: boolean;
  category?: string;
}

export interface UpdateShortcutBody {
  id: string;
  keys: string;
  enabled: boolean;
}

export type GetDashboardActivityParams = {
  limit?: number;
};

export type ListCommandsParams = {
  category?: string;
  search?: string;
};

export type ListHistoryParams = {
  page?: number;
  limit?: number;
  search?: string;
  type?: string;
  status?: string;
};