# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FlowDesk API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemStatus } from "./activityItemStatus";
import type { ActivityItemType } from "./activityItemType";

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  title: string;
  description?: string;
  status: ActivityItemStatus;
  timestamp: string;
  commandId?: string;
}