# lib/api-zod/src/generated/types/chartDataPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FlowDesk API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChartDataPoint {
  date: string;
  executions: number;
  errors: number;
  avgTime: number;
}