# lib/api-zod/src/generated/types/executionResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FlowDesk API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExecutionResultStatus } from "./executionResultStatus";

export interface ExecutionResult {
  id: string;
  commandId: string;
  status: ExecutionResultStatus;
  output?: string;
  error?: string;
  executionTime: number;
  executedAt: string;
}