# lib/api-zod/src/generated/types/historyEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FlowDesk API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HistoryEntryStatus } from "./historyEntryStatus";
import type { HistoryEntryTriggeredBy } from "./historyEntryTriggeredBy";

export interface HistoryEntry {
  id: string;
  commandId?: string;
  commandName?: string;
  type: string;
  status: HistoryEntryStatus;
  output?: string;
  error?: string;
  executionTime: number;
  triggeredBy: HistoryEntryTriggeredBy;
  timestamp: string;
}