# lib/api-zod/src/generated/types/historySummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FlowDesk API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HistorySummary {
  totalExecutions: number;
  successCount: number;
  errorCount: number;
  warningCount: number;
  mostUsedCommand?: string;
  peakHour?: string;
}