# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FlowDesk API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SettingsTheme } from "./settingsTheme";

export interface Settings {
  theme: SettingsTheme;
  language?: string;
  notifications: boolean;
  autoStart: boolean;
  minimizeToTray?: boolean;
  showInDock?: boolean;
  globalShortcutEnabled: boolean;
  clipboardHistoryLimit?: number;
  defaultCategory?: string;
  analyticsEnabled?: boolean;
}