# lib/api-zod/src/generated/types/shortcut.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FlowDesk API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Shortcut {
  id: string;
  name: string;
  description?: string;
  keys: string;
  action: string;
  enabled: boolean;
  category?: string;
}