# lib/api-zod/src/generated/types/updateCommandBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FlowDesk API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateCommandBodyStatus } from "./updateCommandBodyStatus";

export interface UpdateCommandBody {
  name?: string;
  description?: string;
  category?: string;
  trigger?: string;
  script?: string;
  shortcut?: string;
  status?: UpdateCommandBodyStatus;
  tags?: string[];
}